requirements.txt
azure-functions
azure-identity
azure-keyvault
requests
cffi
azure-keyvault-secrets